serving sb-util data from express server
