+++ examples/datavis/index.js
@@ -1,17 +1,23 @@
 const express = require('express');
 const app = express();
-const port = 3000
+const port = 3000;
+const { loadProjectJson, BlockShapes } = require('sb-util');
 
 app.get('/', (req, res) => res.send('Hello World!'));
 
-app.get('/shapes', (req, res) => {
-    res.send([ { name: 'hat', count: 8 },
-    { name: 'boolean', count: 43 },
-    { name: 'reporter', count: 56 },
-    { name: 'c', count: 19 },
-    { name: 'cap', count: 7 },
-    { name: 'stack', count: 40 },
-    { name: 'custom', count: 0 } ]);
+app.get('/shapes', async (req, res) => {
+    const sp = await loadProjectJson('../../tests/data/accelerator.json');
+
+    const blocks = sp.blocks();
+
+    let shapeCounts = [];
+    for (let [key, shape] of Object.entries(BlockShapes)) {
+        const shapeBlocks = blocks.query(`:${shape}`);
+        const blocksCount = Array.from(shapeBlocks.propsIterable()).length;
+        shapeCounts.push({ name: shape, count: blocksCount});
+    }
+
+    res.send(shapeCounts);
 });
 
 app.listen(port, () => console.log(`Example app listening on port ${port}!`))
