+++ examples/datavis/package.json
@@ -9,6 +9,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "express": "^4.17.1"
+    "express": "^4.17.1",
+    "sb-util": "0.0.1-prealpha.4"
   }
 }
