--- examples/node/README.md
@@ -1,4 +1,0 @@
-# sb-util node script
-```
-node index.js
-```
